refactor(#5): 공통 컴포넌트 레이아웃으로 이동
