+++ src/app/[eventId]/match/page.tsx
@@ -1,0 +1,3 @@
+const page = () => <div />;
+
+export default page;
